+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `Moar.DateTime.add` can add a `Moar.Duration` to a `DateTime`
 - `Moar.Retry.rescue_for!/2` can take a `Moar.Duration`
+- Add `Moar.Retry.retry_for` and `Moar.Retry.retry_until`
 
 ## 1.2.0